refactor(eureka): Bump kork; remove keel-eureka
