+++ gradle.properties
@@ -1,6 +1,6 @@
 fiatVersion=1.21.0
 kapt.use.worker.api=true
-korkVersion=7.50.0
+korkVersion=7.51.2
 liquibaseTaskPrefix=liquibase
 org.gradle.parallel=true
 spinnakerGradleVersion=8.3.0
