--- keel-core/src/main/kotlin/com/netflix/spinnaker/keel/info/InstanceIdSupplier.kt
@@ -1,5 +1,0 @@
-package com.netflix.spinnaker.keel.info
-
-import java.util.function.Supplier
-
-interface InstanceIdSupplier : Supplier<String>
