+++ settings.gradle.kts
@@ -36,7 +36,6 @@ include(
   "keel-ec2-api",
   "keel-ec2-plugin",
   "keel-echo",
-  "keel-eureka",
   "keel-front50",
   "keel-igor",
   "keel-orca",
